form/requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
dj-database-url==0.5.0
Django==3.0.1
django-filter==2.2.0
django-phonenumber-field==5.0.0
django-rest-knox==4.1.0
django-suit==0.3a1
django-twilio==0.13.1.post0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
httpie==2.4.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
phonenumbers==8.12.19
Pillow==8.1.2
pycparser==2.20
Pygments==2.8.0
PyJWT==1.7.1
pylint==2.4.4
PyPDF2==1.26.0
PySocks==1.7.1
pytz==2019.3
reportlab==3.5.65
requests==2.25.1
requests-toolbelt==0.9.1
six==1.14.0
sqlparse==0.3.1
twilio==6.54.0
typed-ast==1.4.1
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.11.2
gunicorn==20.0.4
psycopg2-binary==2.8.6

